+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - `all` command added - Creates changelog for markdown, rst and json at the same time.
+- `deprecate` added in `utils.py`.
 
     
 ## [1.0.6] - 2018-03-01